refactor some stats stuff
