+++ Modules/core/Model/CoreInstall.php
@@ -5,6 +5,8 @@ require_once 'Framework/Configuration.php';
 require_once 'Framework/FCache.php';
 require_once 'Framework/Errors.php';
 require_once 'Framework/Statistics.php';
+require_once 'Framework/Events.php';
+
 
 require_once 'Modules/core/Model/CoreStatus.php';
 require_once 'Modules/core/Model/CoreUser.php';
@@ -64,6 +66,22 @@ class CoreDB extends Model {
                 $cp->setShortname($space['id'], $shortname);
             }
         }
+
+        Configuration::getLogger()->debug("[stats] import stats");
+        $cp = new CoreSpace();
+        $statHandler = new EventHandler();
+        $spaces = $cp->getSpaces('id');
+        foreach ($spaces as $space) {
+            $statHandler->spaceCreate(['space' => ['id' => $space['id']]]);
+            $spaceUsers = $cp->getUsers($space['id']);
+            foreach ($spaceUsers as $spaceUser) {
+                $statHandler->spaceUserJoin([
+                    'space' => ['id' => $space['id']],
+                    'user' => ['id' => $spaceUser['id']]
+                ]);
+            }
+        }
+
     }
 
     /**
