+++ Modules/core/Model/CoreSpace.php
@@ -299,7 +299,7 @@ class CoreSpace extends Model {
     public function countSpaces() {
         $sql = "SELECT count(*) FROM core_spaces";
         $res = $this->runRequest($sql)->fetch();
-        return $res[0];
+        return intval($res[0]);
     }
 
     public function setSpace($id, $name, $status, $color, $shortname, $support, $contact) {
@@ -421,7 +421,7 @@ class CoreSpace extends Model {
     public function countUsers($id_space) {
         $sql = "SELECT count(*) FROM core_j_spaces_user WHERE id_space=?";
         $res = $this->runRequest($sql, array($id_space))->fetch();
-        return $res[0];
+        return intval($res[0]);
     }
 
     public function setAdmins($id, $id_admins) {
